chore(identity-auth): move `IdentityAuth` from `libs` to `identity-auth`
