+++ .changeset/heavy-lies-remain.md
@@ -1,5 +1,5 @@
 ---
-"@guardian/libs": minor
+'@guardian/identity-auth': minor
 ---
 
-Adds `IdentityAuth` library for authenticating and authorising users within the Guardian's identity platform hosted by Okta using OAuth2 and OpenID Connect within a client-side, browser-based context.
+Pre-release: Adds `IdentityAuth` library for authenticating and authorising users within the Guardian's identity platform hosted by Okta using OAuth2 and OpenID Connect within a client-side, browser-based context.
