+++ libs/@guardian/identity-auth/src/@types/OAuth.ts
