+++ libs/@guardian/identity-auth/src/authState.ts
