+++ libs/@guardian/identity-auth/src/emitter.ts
