+++ libs/@guardian/identity-auth/src/index.ts
@@ -1,3 +1,21 @@
-const a = 'a';
+import type { IdentityAuthOptions, IdentityAuthState } from './@types/OAuth';
+import type {
+	AccessToken,
+	AccessTokenClaims,
+	CustomClaims,
+	IDToken,
+	IDTokenClaims,
+	Tokens,
+} from './@types/Token';
+import { IdentityAuth } from './identityAuth';
 
-export { a };
+export { IdentityAuth };
+export type { IdentityAuthOptions, IdentityAuthState };
+export type {
+	AccessToken,
+	AccessTokenClaims,
+	CustomClaims,
+	IDToken,
+	IDTokenClaims,
+	Tokens,
+};
