+++ libs/@guardian/libs/src/index.test.ts
@@ -8,8 +8,6 @@ describe('The package', () => {
 			'ArticleElementRole',
 			'ArticlePillar',
 			'ArticleSpecial',
-			'IdentityAuth',
-			'OAuthError',
 			'countries',
 			'debug',
 			'getCookie',
@@ -49,12 +47,6 @@ export type {
 	OphanComponentType,
 	OphanProduct,
 	Switches,
-	IdentityAuthOptions,
-	IdentityAuthState,
-	AccessToken,
-	AccessTokenClaims,
-	IDToken,
-	IDTokenClaims,
 } from './index';
 
 // @ts-expect-error: make sure the above list are real exports
